Rename: assert.h => macros.h
