+++ include/genn_utils/connectors.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // BoB robotics includes
-#include "../common/assert.h"
+#include "../common/macros.h"
 
 // Standard C++ includes
 #include <algorithm>
