+++ include/navigation/image_database.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/pose.h"
 
 // Third-party includes
