+++ include/navigation/insilico_rotater.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 
 // Third-party includes
 #include "third_party/units.h"
