+++ include/navigation/perfect_memory_store_hog.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/logging.h"
 #include "differencers.h"
 #include "ridf_processors.h"
