+++ include/robots/simulated_tank.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/pose.h"
 #include "common/stopwatch.h"
 #include "tank.h"
