+++ include/vicon/udp.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/circstat.h"
 #include "common/logging.h"
 #include "common/pose.h"
