+++ include/viz/sfml_world/sfml_world.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/pose.h"
 #include "robots/robot.h"
