+++ src/common/background_exception_catcher.cc
@@ -1,5 +1,5 @@
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/background_exception_catcher.h"
 
 // Standard C includes
