+++ src/imgproc/opencv_unwrap_360.cc
@@ -1,5 +1,5 @@
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/logging.h"
 #include "imgproc/opencv_unwrap_360.h"
