+++ src/robots/tank.cc
@@ -1,5 +1,5 @@
 // BoB robotics includes
-#include "common/assert.h"
+#include "common/macros.h"
 #include "common/circstat.h"
 #include "common/logging.h"
 #include "robots/tank.h"
